requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
scikit-learn==1.3.2
numpy==1.24.3
requests==2.31.0
google-genai>=1.0.0
groq>=0.9.0
spacy>=3.7.0
python-dotenv>=1.0.0
